==> 3. Descriptive statistics plots.ipynb <==
# Import the necessary packages
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import datetime
from datetime import datetime, timedelta


# Importing the csv file created with the data preparation jupyter notebook.
_2012 = pd.read_csv('2012.csv', sep= ',', low_memory = False, lineterminator='\n')
_2013 = pd.read_csv('2013.csv', sep= ',', low_memory = False, lineterminator='\n')
_2014 = pd.read_csv('2014.csv', sep= ',', low_memory = False, lineterminator='\n')
_2015 = pd.read_csv('2015.csv', sep= ',', low_memory = False, lineterminator='\n')
_2016 = pd.read_csv('2016.csv', sep= ',', low_memory = False, lineterminator='\n')
_2017 = pd.read_csv('2017.csv', sep= ',', low_memory = False, lineterminator='\n')
_2018 = pd.read_csv('2018.csv', sep= ',', low_memory = False, lineterminator='\n')

tweetsPerYear = [
    [2012, len(_2012)],
    [2013, len(_2013)],
    [2014, len(_2014)],
    [2015, len(_2015)],
    [2016, len(_2016)],
    [2017, len(_2017)],
    [2018, len(_2018)]
]
years = [2012, 2013, 2014, 2015, 2016, 2017, 2018]
tweets = [len(_2012),len(_2013),len(_2014),len(_2015),len(_2016),len(_2017),len(_2018)]

flickYear = [
    [2012, 39607],
    [2013, 29614],
    [2014, 34281],
    [2015, 23794],
    [2016, 21660],
    [2017, 19107],
    [2018, 22575]
]
years = [2012, 2013, 2014, 2015, 2016, 2017, 2018]
posts = [39607,29614,34281,23794,21660,19107,22575]

df = pd.DataFrame.from_records(flickYear)


df.rename(columns={0:'Year', 1:'Number of Posts'}, inplace = True)

# Plot the number of tweets per month --> enormous increase.

%matplotlib inline
plt.figure(figsize=(10,7))
sns.set(style='darkgrid')
sns.lineplot(x='Year', y='Number of Posts', markers=True, data=df)
#, bbox_to_anchor=(1, 0.5), ncol=1
plt.savefig('Images/TwitterYearlyActivity.png')


count = 0

for i in posts:
    
    count += i

print(count)

import numpy as np; np.random.seed(10)
import seaborn as sns; sns.set(color_codes=True)
mean, cov = [0, 2], [(1, .5), (.5, 1)]
x, y = np.random.multivariate_normal(mean, cov, size=50).T
ax = sns.kdeplot(x,bw=.15)

plt.figure(figsize=(15, 7))
ax = sns.kdeplot(x,bw=.05, shade=True)
ax = sns.kdeplot(x,bw=.70, shade=True, color="r")
ax.set(xlabel='', ylabel='Density')
plt.savefig('Images/KDE_Example.png')


Data = pd.read_csv('Flickr2014.csv', sep= ',', low_memory = False, lineterminator='\n')

# Descriptive statistics about unique users and tourists for Twitter

# Tourists dataset
Tourists = pd.read_csv('CSV/tourists_total.csv', sep= ',', low_memory = False, lineterminator='\n')

listOfLang = list(Tourists['lang'].value_counts())

total = 0

for i in listOfLang:
    total += i

Tourists['lang'].value_counts()

40315 / total

24868 / total

0.46451739275714665 + 0.2865340077659611

Tourists['tweet_date'] = pd.to_datetime(Tourists['created_at_x'])

#### Further inspect peak around december january

Jan = Tourists[Tourists['tweet_date'].dt.month < 2]

Dec = Tourists[Tourists['tweet_date'].dt.month > 11]

frames = [Jan, Dec]
JanDec = pd.concat(frames)

NYEJan = Jan[Jan['tweet_date'].dt.day < 3]

NYEDec = Dec[Dec['tweet_date'].dt.day > 20]

frames = [NYEJan, NYEDec]
NYEChristmas = pd.concat(frames)

len(NYEChristmas) / len(JanDec)

T2012 = Tourists[Tourists['tweet_date'].dt.year == 2012]
T2013 = Tourists[Tourists['tweet_date'].dt.year == 2013]
T2014 = Tourists[Tourists['tweet_date'].dt.year == 2014]
T2015 = Tourists[Tourists['tweet_date'].dt.year == 2015]
T2016 = Tourists[Tourists['tweet_date'].dt.year == 2016]
T2017 = Tourists[Tourists['tweet_date'].dt.year == 2017]
T2018 = Tourists[Tourists['tweet_date'].dt.year == 2018]

T2018['user_id'].nunique()

uniqueUsers = [
    [2012,20057],
    [2013,24331],
    [2014,24202],
    [2015,14770],
    [2016,10615],
    [2017, 9390],
    [2018, 7066]
]

uniqueTourist = [
    [2012,4351],
    [2013,5297],
    [2014,4998],
    [2015,2533],
    [2016,1385],
    [2017,1061],
    [2018,784]
]

users = pd.DataFrame.from_records(uniqueUsers)
tourists = pd.DataFrame.from_records(uniqueTourist)

users.rename(columns={0:'Year', 1:'Unique Users'}, inplace = True)
tourists.rename(columns={0:'Year', 1:'Unique Users'}, inplace = True)

tables = [users, tourists]
total = pd.concat(tables)

def nameTourists(x):
    
    name = ''
    if x < 6000:
        name = 'Tourists'
    else:
        name = 'Total'
    return name
        

# Create new  column
total[''] = total['Unique Users'].map(lambda x: nameTourists(x))

total

# Plot the number of tweets per month --> enormous increase.
%matplotlib inline
sns.set(style='darkgrid')
sns.lineplot(x='Year', y='Unique Users', markers=["o"], data=users, legend='brief')
sns.lineplot(x='Year', y='Unique Users', markers=["o"], data=tourists, legend='brief')
plt.xlabel("Year") # label for x-axis
plt.ylabel("Number of unique users") # label for y-axis
plt.show()

chart = sns.lineplot(x="Year",
                         y="Unique Users",
                         hue='',
                         data=total
                         )
plt.show()

# Descriptive statistics about unique users and tourists flickr

# Flickr dataset
FlickrTotal = pd.read_csv('FlickrTotal.csv', sep= ',', low_memory = False, lineterminator='\n')
FlickrTourists = pd.read_csv('Flickr_Tourists.csv', sep= ',', low_memory = False, lineterminator='\n')

FlickrTotal['postedTime'] = FlickrTotal['postedDate'].map(lambda x: datetime.fromtimestamp(x))
FlickrTourists['postedTime'] = FlickrTourists['postedDate'].map(lambda x: datetime.fromtimestamp(x))
FlickrTourists['postedTime'] = pd.to_datetime(FlickrTourists['postedTime'])

Flickr2012 = FlickrTotal[FlickrTotal['postedTime'].dt.year == 2012]
Flickr2013 = FlickrTotal[FlickrTotal['postedTime'].dt.year == 2013]
Flickr2014 = FlickrTotal[FlickrTotal['postedTime'].dt.year == 2014]
Flickr2015 = FlickrTotal[FlickrTotal['postedTime'].dt.year == 2015]
Flickr2016 = FlickrTotal[FlickrTotal['postedTime'].dt.year == 2016]
Flickr2017 = FlickrTotal[FlickrTotal['postedTime'].dt.year == 2017]
Flickr2018 = FlickrTotal[FlickrTotal['postedTime'].dt.year == 2018]

FlickrT2012 = FlickrTourists[FlickrTourists['postedTime'].dt.year == 2012]
FlickrT2013 = FlickrTourists[FlickrTourists['postedTime'].dt.year == 2013]
FlickrT2014 = FlickrTourists[FlickrTourists['postedTime'].dt.year == 2014]
FlickrT2015 = FlickrTourists[FlickrTourists['postedTime'].dt.year == 2015]
FlickrT2016 = FlickrTourists[FlickrTourists['postedTime'].dt.year == 2016]
FlickrT2017 = FlickrTourists[FlickrTourists['postedTime'].dt.year == 2017]
FlickrT2018 = FlickrTourists[FlickrTourists['postedTime'].dt.year == 2018]



FlickrT2012['userID'].nunique()

uniqueFlickrUsers = [
    [2012,1233],
    [2013,1190],
    [2014,1111],
    [2015,1130],
    [2016,1087],
    [2017, 904],
    [2018, 740]
]

uniqueFlickrTourist = [
    [2012,276],
    [2013,238],
    [2014,201],
    [2015,235],
    [2016,191],
    [2017,189],
    [2018,143]
]

count = 0

for i in uniqueFlickrTourist:
    count += i[1]
print(count)

#### Post distribution over time/week/year

# Create two new columns, one for tweet time the other for month.
Tourists['tweet_hour'] = Tourists['tweet_date'].dt.hour
Tourists['tweet_month'] = Tourists['tweet_date'].dt.month
# Same for flickr
FlickrTourists['post_hour'] = FlickrTourists['postedTime'].dt.hour
FlickrTourists['post_month'] = FlickrTourists['postedTime'].dt.month


# Get tweets per month
TweetsPerMonth = Tourists.groupby('tweet_month', as_index=False).count()
# Get flickr posts per month
FlickrPerMonth = FlickrTourists.groupby('post_month', as_index=False).count()

TweetsPerMonth

countFlickr = 0
for i in list(FlickrPerMonth['latitude']):
    
    countFlickr += i
print(countFlickr)

countTwitter = 0
for i in list(TweetsPerMonth['item_number']):
    
    countTwitter += i
print(countTwitter)

relativeTweetsPerMonth = []

for i in list(TweetsPerMonth['item_number']):
    
    relativeTweetsPerMonth.append(i/117777)

relativeFlickrPerMonth = []

for i in list(FlickrPerMonth['latitude']):
    
    relativeFlickrPerMonth.append(i/27806)

FlickrMonthly = {"Month":[1,2,3,4,5,6,7,8,9,10,11,12], "Relative number of Tweets": relativeFlickrPerMonth, "Platform": [
    "Flickr","Flickr","Flickr","Flickr","Flickr","Flickr","Flickr","Flickr","Flickr","Flickr","Flickr","Flickr"]
}

TwitterMonthly = {"Month":[1,2,3,4,5,6,7,8,9,10,11,12], "Relative number of Tweets": relativeTweetsPerMonth, "Platform": [
    "Twitter","Twitter","Twitter","Twitter","Twitter","Twitter","Twitter","Twitter","Twitter","Twitter","Twitter","Twitter"]
}

# Plot the number of tweets per month --> enormous increase.
%matplotlib inline
sns.set(style='darkgrid')
sns.lineplot(x='Year', y='Unique Users', markers=["o"], data=users, legend='brief')
sns.lineplot(x='Year', y='Unique Users', markers=["o"], data=tourists, legend='brief')
plt.xlabel("Year") # label for x-axis
plt.ylabel("Number of unique users") # label for y-axis
plt.show()

RelativeMonthlyPosts = pd.DataFrame(TwitterMonthly) 

RelativeMonthlyPostsFlickr = pd.DataFrame(FlickrMonthly) 

RelativeMonthlyPosts = RelativeMonthlyPosts.append(RelativeMonthlyPostsFlickr)

RelativeMonthlyPosts = RelativeMonthlyPosts.rename(columns={'Relative number of Tweets':"Relative number of posts"})

plt.figure(figsize=(15,7))
sns.set_style('darkgrid')
chart = sns.lineplot(x="Month",
                         y="Relative number of posts",
                         hue='Platform',
                         data=RelativeMonthlyPosts
                         )

chart.legend(loc='upper left', fancybox=True, framealpha=1, shadow=True, borderpad=1)
plt.setp(chart.get_legend().get_texts(), fontsize='15') # for legend text
plt.setp(chart.get_legend().get_title(), fontsize='10') # for legend title
plt.rcParams["axes.labelsize"] = 15
#, bbox_to_anchor=(1, 0.5), ncol=1
plt.savefig('Images/relativeMonthlyPosts.png')



# Get tweets per hour
TweetsPerHour = Tourists.groupby('tweet_hour', as_index=False).count()
# Get flickr posts per month
FlickrPerHour = FlickrTourists.groupby('post_hour', as_index=False).count()

countHourFlickr = 0
for i in list(FlickrPerHour['latitude']):
    
    countHourFlickr += i
print(countHourFlickr)

countHourTwitter = 0
for i in list(TweetsPerHour['item_number']):
    
    countHourTwitter += i
print(countHourTwitter)

relativeTweetsPerHour = []

for i in range(len(list(TweetsPerHour['item_number']))):
    if i == 0:
        relativeTweetsPerHour.append(TweetsPerHour['item_number'][23]/117777)
    if i != 23:
        relativeTweetsPerHour.append(TweetsPerHour['item_number'][i]/117777)

relativeTweetsPerHour

relativeFlickrPerHour = []

for i in range(len(list(FlickrPerHour['latitude']))):
    if i == 0:
        relativeFlickrPerHour.append(FlickrPerHour['latitude'][23]/27806)
    if i != 23:
        relativeFlickrPerHour.append(FlickrPerHour['latitude'][i]/27806)

relativeFlickrPerHour

FlickrHourly = {"Hour":[0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23], "Relative number of Posts": relativeFlickrPerHour, "Platform": [
    "Flickr","Flickr","Flickr","Flickr","Flickr","Flickr","Flickr","Flickr","Flickr","Flickr","Flickr","Flickr",
"Flickr","Flickr","Flickr","Flickr","Flickr","Flickr","Flickr","Flickr","Flickr","Flickr","Flickr","Flickr"]
}

TwitterHourly = {"Hour":[0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23], "Relative number of Posts": relativeTweetsPerHour, "Platform": [
    "Twitter","Twitter","Twitter","Twitter","Twitter","Twitter","Twitter","Twitter","Twitter","Twitter","Twitter","Twitter",
"Twitter","Twitter","Twitter","Twitter","Twitter","Twitter","Twitter","Twitter","Twitter","Twitter","Twitter","Twitter"]
}

RelativeHourlyPosts = pd.DataFrame(TwitterHourly) 

RelativeHourlyPostsFl = pd.DataFrame(FlickrHourly) 

RelativeHourlyPosts = RelativeHourlyPosts.append(RelativeHourlyPostsFl)

plt.figure(figsize=(15,7))
sns.set_style('darkgrid')
chart = sns.lineplot(x="Hour",
                         y="Relative number of Posts",
                         hue='Platform',
                         data=RelativeHourlyPosts
                         )

chart.legend(loc='upper left', fancybox=True, framealpha=1, shadow=True, borderpad=1)
plt.setp(chart.get_legend().get_texts(), fontsize='15') # for legend text
plt.setp(chart.get_legend().get_title(), fontsize='10') # for legend title

#, bbox_to_anchor=(1, 0.5), ncol=1
plt.savefig('Images/relativeHourlyPosts.png')



# Some stats about the average number of posts, time between posts etc

Tourists.head(2)

print('unique users: {}'.format(Tourists['user_id'].nunique()))
print('tweets per user: {}'.format(len(Tourists) / Tourists['user_id'].nunique()))

print('unique users: {}'.format(FlickrTourists['userID'].nunique()))
print('flickr posts per user: {}'.format(len(FlickrTourists) / FlickrTourists['userID'].nunique()))

#### Get average time between first and last posts

# Get all the unique users and put them in a list
UniqueUsers = FlickrTourists['userID'].unique()

FlickrFirstLast = []

FlickrFirstLast

# In this loop, get the average time between first and last post
for i in UniqueUsers:
    
    
    # Get all the tweets for this particular user.
    TweetsForThisUser = FlickrTourists[FlickrTourists['userID'] == i]
    numberOfTweets = len(TweetsForThisUser)
    
    # Only continue when there are more than 1 tweets.
    if numberOfTweets > 1:
        
        # Sort the tweets and get the first and last tweet of a user.
        TweetsForThisUser = TweetsForThisUser.sort_values(by='postedTime')
        FirstAndLast = TweetsForThisUser.iloc[[0, -1]]
        
        # Get the difference in days between the first and last tweet.
        secondsDelta = (FirstAndLast['postedTime'].iloc[0] - FirstAndLast['postedTime'].iloc[1]).seconds
        
        FlickrFirstLast.append([secondsDelta, numberOfTweets])
        
    else:
        continue

FlickrTimeBetweenFirstLast = []

for i in FlickrFirstLast:
    FlickrTimeBetweenFirstLast.append(i[0])

len(FlickrTimeBetweenFirstLast)

count = 0

for i in FlickrTimeBetweenFirstLast:
    count += i
    
count = count / 1471

count

intervals = (
    ('weeks', 604800),  # 60 * 60 * 24 * 7
    ('days', 86400),    # 60 * 60 * 24
    ('hours', 3600),    # 60 * 60
    ('minutes', 60),
    ('seconds', 1),
    )

def display_time(seconds, granularity=2):
    result = []

    for name, count in intervals:
        value = seconds // count
        if value:
            seconds -= value * count
            if value == 1:
                name = name.rstrip('s')
            result.append("{} {}".format(value, name))
    return ', '.join(result[:granularity])

display_time(count)

TwitterFirstLast

FlickrTimeBetweenPosts = []

for i in FlickrFirstLast:
    FlickrTimeBetweenPosts.append(i[0]/i[1])

tbetween = 0

for i in FlickrTimeBetweenPosts:
    tbetween += i

tbetween = tbetween/1471


display_time(tbetween)

FlickrTimeBetweenPosts

# Get some stats about number of titles and number of descriptions

TouristsFlickr = pd.read_csv('Flickr_Tourists.csv', sep= ',', low_memory = False, lineterminator='\n')

TouristsFlickr.shape

# Tourists dataset
Tourists = pd.read_csv('tourists_total.csv', sep= ',', low_memory = False, lineterminator='\n')

Tourists.shape

FlickrDescription = TouristsFlickr[TouristsFlickr['description'].isnull() == False]

TouristsFlickr['title'].isnull().value_counts()

Tourists[Tourists['text'] == ' ']

FlickrLDA = FlickrDescription[FlickrDescription['title'].isnull() == False]

# Create a list to store the average number of words in
NumberOfWords = []

# Iterate over FlickrLDA to get the average number of words
for index, row in FlickrLDA.iterrows():
    numberOfWords = 0
    
    titleCount = row['title'].count(' ')
    descriptionCount = row['description'].count(' ')
    numberOfWords = titleCount + descriptionCount
    
    # Add to list
    NumberOfWords.append(numberOfWords)

total = 0
for i in NumberOfWords:
    total += i

total

total / len(NumberOfWords)

# Create a list to store the average number of words in
NumberOfWordsT = []

# Iterate over FlickrLDA to get the average number of words
for index, row in Tourists.iterrows():
    numberOfWords = row['text'].count(' ')
    
    # Add to list
    NumberOfWordsT.append(numberOfWords)

totalT = 0
for i in NumberOfWordsT:
    totalT += i

totalT

totalT / len(NumberOfWordsT)

len(NumberOfWordsT)

FlickrLDA.to_csv('FlickrLDA.csv', index=False)

# Hashtag dataset
HashtagData = pd.read_csv('TwitterHashtags_LDA.csv', sep= ',', low_memory = False, lineterminator='\n')

# Create a list to store the average number of words in
NumberOfWords = []

# Iterate over FlickrLDA to get the average number of words
for index, row in HashtagData.iterrows():
    numberOfWords = 0
    
    titleCount = row['title'].count(' ')
    descriptionCount = 
    numberOfWords = row['description'].count(' ')
    
    # Add to list
    NumberOfWords.append(numberOfWords)
