==> 4C. Aggregating themes of LDA.ipynb <==
# Import the necessary packages
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import datetime
from datetime import datetime, timedelta


# Importing the datasets.
Tourists = pd.read_csv('tourists_total.csv', sep= ',', low_memory = False, lineterminator='\n')
LDA_EN = pd.read_csv('tweets_topic_EN.csv', sep= ',', low_memory = False, lineterminator='\n')
LDA_NL = pd.read_csv('tweets_topic_NL.csv', sep= ',', low_memory = False, lineterminator='\n')

Tourists.head()

LDA_EN.head()

LDA_NL.head()

# Create one dataframe with all topic information.
frames = [LDA_EN, LDA_NL]
LDA_Total = pd.concat(frames)

Tourists_LDA = pd.merge(Tourists, LDA_Total, on='item_number', how='left')

Tourists['item_number'].nunique()

Tourists_LDA.shape

# Drop potential duplicates.
Tourists_LDA.drop_duplicates(subset ="item_number", keep = 'first', inplace = True) 

Tourists_LDA['item_number'].nunique()

# Creating themes based on the topics.
themeDict = {}

# Iterate over the dataframe to select the theme.
for index, row in Tourists_LDA.iterrows():
    
    # If we are dealing with a dutch tweet.
    if row['lang'] == 'nl':
        
        if row['topic'] == 'Dutch Topic 3' or row['topic'] == 'Dutch Topic 7' or row['topic'] == 'Dutch Topic 18':
            
            # Update dict.
            themeDict[row['item_number']] = 'Museums'
        
        elif row['topic'] == 'Dutch Topic 4' or row['topic'] == 'Dutch Topic 6' or row['topic'] == 'Dutch Topic 10' or row['topic'] == 'Dutch Topic 12':
            # Update dict.
            themeDict[row['item_number']] = 'Festivals'
            
        elif row['topic'] == 'Dutch Topic 17' or row['topic'] == 'Dutch Topic 18':
            
            # Update dict.
            themeDict[row['item_number']] = 'Sightseeing'
        
        elif row['topic'] == 'Dutch Topic 8' or row['topic'] == 'Dutch Topic 20':
            
            # Update dict.
            themeDict[row['item_number']] = 'Sports'
        
        elif row['topic'] == 'Dutch Topic 2':
            
            # Update dict.
            themeDict[row['item_number']] = 'Travel'
        
        else:
            continue
    
    # If we are dealing with an english tweet.
    elif row['lang'] == 'en':
        
        if row['topic'] == 'Topic 2':
            
            # Update dict.
            themeDict[row['item_number']] = 'Museums'
        
        elif row['topic'] == 'Topic 1' or row['topic'] == 'Topic 3' or row['topic'] == 'Topic 6':
            
            # Update dict.
            themeDict[row['item_number']] = 'Festivals'
        
        elif row['topic'] == 'Topic 7' or row['topic'] == 'Topic 8' or row['topic'] == 'Topic 11' or row['topic'] == 'Topic 16':
            
            # Update dict.
            themeDict[row['item_number']] = 'Sightseeing'
            
        elif row['topic'] == 'Topic 19':
            
            # Update dict.
            themeDict[row['item_number']] = 'Sports'
        
        elif row['topic'] == 'Topic 13' or row['topic'] == 'Topic 15' or row['topic'] == 'Topic 17':
            
            # Update dict.
            themeDict[row['item_number']] = 'Travel'
        
        else:
            continue
    
    else:
        continue

# Create DF.
themeDF = pd.DataFrame(themeDict, index=[0]).T

# Rename columns
themeDF.rename(columns={0:'theme'}, inplace=True)

themeDF

# Merge with original tourist df.
Tourists_LDA_2 = pd.merge(Tourists_LDA, themeDF, left_on='item_number', right_index= True, how='left')

# Getting some basic descriptive statistics.
ThemeList = list(Tourists_LDA_2_English['theme'].value_counts())

count = 0

for i in ThemeList:
    count += i
count

Tourists_LDA_2_Dutch = Tourists_LDA_2[Tourists_LDA_2['lang'] == 'nl']
Tourists_LDA_2_English = Tourists_LDA_2[Tourists_LDA_2['lang'] == 'en']

Tourists_LDA_2_English['theme'].value_counts()

4677 / 20301

# Write to CSV for later use.
Tourists_LDA_2.to_csv('tourists_total_themes.csv', index=False)
