==> 4D. Creating PhotoTweets Dataset.ipynb <==
# Import the necessary packages
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import matplotlib.patches as mpatches
import seaborn as sns
import sqlite3
sns.set()

# Getting all monthly files containing the information about attachments
# Connect with the db file and create cursor object to be able to execute queries
year = 2012
month = 1

yearStr = ''
monthStr = ''

while year < 2019:
    
    while month < 13:
        
        # Change the month to corresponding string.
        if month == 1:
            monthStr = '01'
        if month == 2:
            monthStr = '02'    
        if month == 3:
            monthStr = '03'
        if month == 4:
            monthStr = '04'
        if month == 5:
            monthStr = '05'
        if month == 6:
            monthStr = '06'
        if month == 7:
            monthStr = '07'    
        if month == 8:
            monthStr = '08'
        if month == 9:
            monthStr = '09'
        if month == 10:
            monthStr = '10'
        if month == 11:
            monthStr = '11'
        if month == 12:
            monthStr = '12'
        
        yearStr = str(year)
        # Create connection
        connection = sqlite3.connect('/Volumes/Hard drive/Thesis/Data/' + yearStr + '/month_' + yearStr + '_' + monthStr +'.db')
        cursor = connection.cursor()
        
        # Read media nl file and store it in the final dataframe.
        media = pd.read_sql_query("SELECT * FROM media_NL", connection)
        TwitterMedia = TwitterMedia.append(media)
        
        # Add one month
        month += 1
        
    # Reset month
    month = 1
    
    # Add one year
    year += 1      
        

# Create media id dataframe
TwitterMedia = pd.DataFrame()

TwitterMedia
TwitterMedia.to_csv('CSV/TwitterPhotoIDs.csv', index=False)

Tourists = pd.read_csv('tourists_total.csv', sep= ',', low_memory = False, lineterminator='\n')
PhotoIDs = pd.read_csv('TwitterPhotoIDs.csv', sep=',', low_memory = False, lineterminator='\n')

Tourists.head()

PhotoIDs.head()

tweetsWithPhoto = list(PhotoIDs['item_number'])

len(tweetsWithPhoto)

# Create an empty dictionary to whether a tweet has photos
PhotoDict = dict()

for index, row in Tourists.iterrows():
    
    hasPhoto = False
    
    if row['item_number'] in tweetsWithPhoto:
        hasPhoto = True
    
    PhotoDict.update( {row['item_number'] : hasPhoto} )

# Transform the dictionary into a dataframe and rename the columns.
PhotoInfo = pd.DataFrame(list(PhotoDict.items()))
PhotoInfo = PhotoInfo.rename(columns={0: "item_number", 1: "has_attachment"})

# Statistics on the amount of tweets with attachments
PhotoInfo['has_attachment'].value_counts()

# Merge with the original dataset
Tourists = pd.merge(Tourists, PhotoInfo, on='item_number', how='left')

# Export to CSV files for later analysis.
Tourists.to_csv('tourists_total.csv', index=False)

PhotoTourists = Tourists[Tourists['has_attachment'] == True]

PhotoTourists['has_attachment'].value_counts()

# Export to CSV files for later analysis.
PhotoTourists.to_csv('tourists_attachments.csv', index=False)

Tourists_Attach = pd.read_csv('tourists_attachments.csv', sep= ',', low_memory = False, lineterminator='\n')
